Fix display of indented text in PowerPoint.
PowerPoint seems to collapse whitespace in a super-odd way: single spaces
are preserved, TWO spaces are preserved, but runs of three or more spaces
are condensed down to two spaces. Wha?? Luckily, it respects nbsp.

To avoid making lines impossible to wrap, we only replace every *other*
space with an nbsp. This is still enough to keep PPT happy.